moved school with relation to separate file
